Project Launch I and II
-PLEASE READ


🔥 🔥 🔥 🔥 🔥 🔥 🔥


Git Hub Community, I am an aspiring entrepreneur, 27, on his nth venture. N is greater than 30. It has been an arduous journey, but I have learned many lessons. Finally putting it together with the realization I need help. 

Project I: 
Background involves an actionable plan which has potential to extract monopoly profits from an incredibly niche industry. I sold my car to file a provisional patent (hardware/software integration) and have lost nearly all my friends in order to maintain a state of neurotic creativity to keep the dream alive. I cannot go to market if I do not have a professionalized digital experience that can attract user adoption quickly which simultaneously prevents the 1,000,000 questions or doubts which will come up if that experience is not simplified. Even then, when all questions are ans
red there is rejection because of an ideological gap or the hesitation time creates. This has revealed to me the close-mindedness of human nature which is why a pleasant user experience which gravitationally pulls users in is the only way to achieve this adoption. I am simply attempting to give my life to something productive and beneficial. It is very frustrating when you are unable to do so and are constantly questioned your intentions.

What needs to be built:
What we seek to build is a 3 page front-end experience in Django and Integrated back-end in Python (OPEN TO ALL INPUT. My research suggest this is best?)
The "great-white" has ripped me off countless ventures (at nearly every period in the venture cycle). Due to this, we desire to have a strong entry and consent process. What I mean by this: "click to enter site" = massive scroll (which speeds the "reading" up / equivalent to facial recognition) which prevents any IP theft on the platform. I desire to have the consent sync with the user's phone contacts and email addresses for assurance. Consent would include no ability to distribute, not compete within industry for 20+ years, inability to create a similar site etc. If you have sold a start-up or app maybe you have proper legal documentation, if not I can track this down. What is more important is the ability to upload legal documentation with a pleasant scroll feature. Platform needs to come first. 

A pleasant experience I have discovered exist on Tesla's Career Page: https://www.tesla.com/careers/search/job/apply/132673

While the information gathered and subject titles will be different the experience is pleasant. I would list what I desire in terms "subject titles"
here but that would compromise the market I am targeting. I also desire the ability to collect multiple customer documents in a drag and drop feature similar to the site. Some front-end subjects will hopefully possess drop-down arrows which have several options. These options will be synced or scraped from a government site (public data). However, we desire the experience to be far superior (easier to navigate, and only possess core data) than what government currently has. Designs are complete with a desired layout which we can go over should you all choose to help me. Lastly, a checkout feature is desired - connecting to stripe or square. We would love for a crypto wallet and checkout capability to be included syncing to an appropriate exchange (one of the larger ones).

While many of you could care less about my personal life circumstances a uniform landing page/user experience would be beneficial
for this entire community. It would also serve to reduce redundancy, while increasing efficiency. The majority of small web-startups are gathering and flipping data.

-Project II: 
This project is fun-oriented and charitable. True Americans love risk (: For these reasons I have created a fin-tech business plan and finished the design process for a web-application and corresponding app.  Broadly speaking, I believe the black community is underserved by the investment community. In fact, I believe it is highly probable that it is intentional. As we all know, without 51% of the collective unconscious holding a uniform belief, my belief will remain "conspiracy." My kicking and screaming has not achieved the results I have desired regarding this issue, so I have come to GitHub.
Nonetheless, the securities Act was written in 1933. Long before the advent of the internet. Modern securities laws disallow disenfranchised individuals to take a chance on a limited partnership, or private investment opportunity without being an "accredited investor." I would argue equality of an investor has been substantially liberalized since 1933 and that lottery tickets should have far more regulations than private investment opportunities or stock picking. We have finally completed a lite-paper which involves increasing the elasticity of supply of the hedge-fund industry and I possess a value algorithm in python which has before well. We chose 7 stocks and 2 have been taken out or were acquired on wall-street in take-privates. Within 24 months of releasing our paper on Linkedin. There have a few several start-ups which were launched with a similar premise. Titan is a good example of one. Titan does not charge the 2-20 split and lowers the amount of capital required for an investment. However, their funds are not actively managed by HF managers. Titan tracks the portfolios top managers (presumably through 13-qs) which trail by a quarter. We desire to utilize actively managed "quanta-mental" strategies to manage capital for disenfranchised citizens. The current paradigm requires a form of elitism not just to actively manage funds on wall-street but also to participate. We desire to eliminate both barriers. I do not feel comfortable releasing this lite-paper publicly without proper protection and for that reason...

What needs to be built:
I. A web-application similar to project one. Web application has a similar set up to project one. Same template could serve as a baseline. 3 pages, data collection along with document drag and drop feature to collect information of investors (to deem whether they are accredited) and a checkout feature through stripe, square and connection with a crypto wallet. There are numerous platforms offering this form of service but the user experience is very poor. Less is more on the web in our opinion. Pleasant experience (simple clicks) and hard data analysis/reorganization back-end. An additional feature will include our litepaper and investment philosophy prior to checkout. As review: log-in 1) data/privacy/non-compete/connection to all contacts within phone, 2) data collection 3) agreement not to sue, litepaper and checkout. 4) Side tab for research and performance tracking.

II) corresponding app. Preferably in swift with same features.

******* I am fully committed to Project I but am more than capable of handling both projects. Capital allocation and business building go hand in hand*****
******* I need so much help but the compound effect on others will be substantial******
🔥 🔥 🔥 🔥 🔥 🔥 🔥 🔥 🔥 

design features and drawings may be added later subject to security clearance.
I do not run out of ideas... helping me launch here will be start. 
I happen to be obsessed with Future, the rapper. I also love Lebron James and Tom Brady. In the event we succeed here and I 
achieve some form of wealth creation, I will co-produce What A Time to Be Alive 2 (Album), and Purple Reign 2 (Mixtape)... ENSURING
that the GitHub community gets a lyrical shoutout. Maybe, just maybe, I will get GitHub tattooed (I have none - we can even do it on Lebron's The Shop")